more cleaning up, renaming
